Adicionando handler de criação de jwt
