+++ src/controllers/AuthControllers/logoutUserController.ts
