+++ src/handlers/generateToken.ts
@@ -1,0 +1,12 @@
+import { sign } from 'jsonwebtoken'
+export const generateToken = (
+  userId: string,
+  expiresIn: string,
+  secret: string
+) => {
+  const token = sign({ id: userId }, secret, {
+    subject: userId,
+    expiresIn: expiresIn,
+  })
+  return token
+}
