+++ src/routes/auth.ts
@@ -1,10 +1,10 @@
 import { Router } from 'express'
 
+import login from '../controllers/AuthControllers/loginUserController'
 import { createTokenController } from '../controllers/TokenControllers/createTokenController'
 import { deleteToken } from '../controllers/TokenControllers/deleteTokenController'
 import { getTokenController } from '../controllers/TokenControllers/getTokenController'
 import { updateTokenController } from '../controllers/TokenControllers/updateTokenController'
-import login from '../controllers/UserControllers/loginUserController'
 const router = Router()
 router.post('/login', login)
 router.get('/token', getTokenController)
